Extracted symbols from main.py into their own modules.
